+++ frb_rust/src/lockable/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod base;
 pub(crate) mod implementor;
+mod immediate_invalid_borrow_checker;
