+++ frb_rust/src/rust_auto_opaque/dart2rust_implicit.rs
@@ -1,5 +1,4 @@
 use crate::for_generated::BaseArc;
-use crate::rust_async::{RwLockReadGuard, RwLockWriteGuard};
 use crate::rust_auto_opaque::inner::RustAutoOpaqueInner;
 use crate::rust_auto_opaque::order::RustAutoOpaqueOrder;
 use crate::rust_opaque::RustOpaqueBase;
@@ -61,90 +60,3 @@ pub fn rust_auto_opaque_decode_compute_order(
     sorted_infos.into_iter().map(|info| info.index).collect()
 }
 
-fn check_no_immediate_invalid_borrow(sorted_infos: &[RustAutoOpaqueLockOrderInfo]) -> bool {
-    let mut last_object = None;
-    let mut checker = ImmediateInvalidBorrowChecker::default();
-
-    for info in sorted_infos {
-        if Some(info.object_order) != last_object {
-            last_object = Some(info.object_order);
-            checker = ImmediateInvalidBorrowChecker::default();
-        }
-
-        if !checker.update(info.mutable) {
-            return false;
-        }
-    }
-
-    true
-}
-
-#[derive(Default)]
-struct ImmediateInvalidBorrowChecker {
-    count: usize,
-    has_mutable: bool,
-}
-
-impl ImmediateInvalidBorrowChecker {
-    fn update(&mut self, mutable: bool) -> bool {
-        self.count += 1;
-        self.has_mutable |= mutable;
-        (!self.has_mutable) || (self.count <= 1)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_check_no_immediate_invalid_borrow() {
-        assert!(check_no_immediate_invalid_borrow(&[]));
-
-        for mutable in [false, true] {
-            assert!(check_no_immediate_invalid_borrow(&[
-                RustAutoOpaqueLockOrderInfo {
-                    index: 0,
-                    mutable,
-                    object_order: RustAutoOpaqueOrder::new_for_test(100),
-                }
-            ]));
-        }
-
-        for (mutable_a, mutable_b, expect) in [
-            (false, false, true),
-            (false, true, false),
-            (true, false, false),
-            (true, true, false),
-        ] {
-            assert_eq!(
-                check_no_immediate_invalid_borrow(&[
-                    RustAutoOpaqueLockOrderInfo {
-                        index: 0,
-                        mutable: mutable_a,
-                        object_order: RustAutoOpaqueOrder::new_for_test(100),
-                    },
-                    RustAutoOpaqueLockOrderInfo {
-                        index: 1,
-                        mutable: mutable_b,
-                        object_order: RustAutoOpaqueOrder::new_for_test(100),
-                    }
-                ]),
-                expect
-            );
-        }
-
-        assert!(check_no_immediate_invalid_borrow(&[
-            RustAutoOpaqueLockOrderInfo {
-                index: 0,
-                mutable: true,
-                object_order: RustAutoOpaqueOrder::new_for_test(100),
-            },
-            RustAutoOpaqueLockOrderInfo {
-                index: 1,
-                mutable: true,
-                object_order: RustAutoOpaqueOrder::new_for_test(101),
-            }
-        ]));
-    }
-}
